feat: fencing before start, better cleanup
- fencing service before start
- clean_up renamed
- refactor
- fancier print out
